+++ nummus/controllers/health.py
@@ -12,7 +12,6 @@ import flask
 from nummus import health_checks, web
 from nummus.controllers import base
 from nummus.models import Config, ConfigKey, HealthCheckIssue, YIELD_PER
-from nummus.web import utils as web_utils
 
 
 class _HealthContext(TypedDict):
@@ -69,7 +68,7 @@ def ignore(uri: str) -> str:
     """
     p = web.portfolio
     with p.begin_session() as s:
-        c = web_utils.find(s, HealthCheckIssue, uri)
+        c = base.find(s, HealthCheckIssue, uri)
         c.ignore = True
         name = c.check
